tests: enable kdump.service in the systemd preset policy to allow setting up the crashkernel kernel parameter automatically
Previously, the kernel parameter crashkernel will be set up when
kexec-tools is installed. But now it will happen only when kdump.service
is enabled first. When kexec-tools is freshly installed, the kdump.service
will be enabled according to the systemd policy file and later the
kernel parameter crashkernel will be added only if kdump.service is
enabled.  Fedora disables kdump.service in the systemd policy. So enable
kdump.service in the systemd preset policy manually in the test script.

In case kexec-tools has already been installed, simply enable
kdump.service.

Fixes: 0ffce0ef ("Only try to reset crashkernel when kdump.service is enabled")
Signed-off-by: Coiby Xu <coxu@redhat.com>